pyproject.toml
[project]
name = "x-digest"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "requests-oauthlib",
    "anthropic",
    "jinja2",
    "pyyaml",
]

[tool.uv]
package = false
